Mapping Experiment 2: JSON -> models with Jane-php library and Symfony normalizers
This explores generating PHP classes using the Jane PHP library:

https://jane.readthedocs.io/en/latest/documentation/JsonSchema.html

On the upside, it creates models, normalizers, and validators which
is all we need to go from JSON and Builder objects to CompiledBlueprint.
The validators are statically declared and don't rely on parsing and
validating data usint the JSON schema.

On the downside:
* The generated model classes have an `$initialized` property which we don't need but the normalizers seem to rely on it.
* Validating collections like blueprint.steps doesn't see to be covered
  by the generated validators