Downgrade to PHP 7.4, move from Pest to PHPUnit, establish CI (#25)
- Downgrades the lib to PHP 7.4
- Adds 'Check PHP CS == 7.4' step to GitHub Actions
- Replaces Pest with PHPUnit
- Rewrites RmStep tests in PHPUnit 
- Adds matrix unit tests step for all operating systems and all PHP
versions >= 7.4

 ### Caveats

- Dependencies versions are wildcards. This is inherently unstable, but
our true goal is to have stable 7.0 support. Establishing fixed versions
is a tedious job that does not seem crucial at this temporary step.
- Unit tests should include `prefer-lowest` as one of the arguments in
the `dependency_version` list. This will allow for the broadest version
support. But at this point in time it fails at 7.4. Fixing this serves
little purpose as, see logic above.
- Although we know PHPUnit 6.5 should be used, since this is the newest
versions supporting PHP == 7.0, PHPUnit 9.6 is used here. This is due to
all lower versions throwing the pesky `PHP Fatal error: Cannot acquire
reference to $GLOBALS`. This is still a step in the right direction as
writing PHPUnit tests in 9.6 and then downgrading them to 6.5 is still
easier, than writing in Pest and then going for PHPUnit 6.5.
- It should be analyzed if it is in fact beneficial to run the matrix
with each pull/pr. Alternatives: cron, only pr and more.


Co-authored-by: Adam Zieliński <adam@adamziel.com>